feat(Cmdlets): ✨ Introduce new cmdlets for credential management
* Added `ConvertTo-GraphCredential`, `ConvertTo-OAuth2Credential`, and `ConvertTo-SendGridCredential` cmdlets to create `PSCredential` objects for Microsoft Graph, OAuth2, and SendGrid API authentication respectively.
* Removed obsolete cmdlet implementations for better code organization and clarity.
